netfilter: xt_qtaguid: fix compile warnings
Fix compilation warnings by type-casting atomic_64 types
to unsigned long long

+++ net/netfilter/xt_qtaguid.c
@@ -1957,18 +1957,30 @@ static int qtaguid_ctrl_proc_show(struct seq_file *m, void *v)
 			   "match_found_no_sk_in_ct=%llu "
 			   "match_no_sk=%llu "
 			   "match_no_sk_file=%llu\n",
-			   atomic64_read(&qtu_events.sockets_tagged),
-			   atomic64_read(&qtu_events.sockets_untagged),
-			   atomic64_read(&qtu_events.counter_set_changes),
-			   atomic64_read(&qtu_events.delete_cmds),
-			   atomic64_read(&qtu_events.iface_events),
-			   atomic64_read(&qtu_events.match_calls),
-			   atomic64_read(&qtu_events.match_calls_prepost),
-			   atomic64_read(&qtu_events.match_found_sk),
-			   atomic64_read(&qtu_events.match_found_sk_in_ct),
-			   atomic64_read(&qtu_events.match_found_no_sk_in_ct),
-			   atomic64_read(&qtu_events.match_no_sk),
-			   atomic64_read(&qtu_events.match_no_sk_file));
+			   (unsigned long long)
+			   	atomic64_read(&qtu_events.sockets_tagged),
+			   (unsigned long long)
+				atomic64_read(&qtu_events.sockets_untagged),
+			   (unsigned long long)
+				atomic64_read(&qtu_events.counter_set_changes),
+			   (unsigned long long)
+				atomic64_read(&qtu_events.delete_cmds),
+			   (unsigned long long)
+			   	atomic64_read(&qtu_events.iface_events),
+			   (unsigned long long)
+			   	atomic64_read(&qtu_events.match_calls),
+			   (unsigned long long)
+			   	atomic64_read(&qtu_events.match_calls_prepost),
+		           (unsigned long long)
+			   	atomic64_read(&qtu_events.match_found_sk),
+			   (unsigned long long)
+			   	atomic64_read(&qtu_events.match_found_sk_in_ct),
+			   (unsigned long long)
+			   	atomic64_read(&qtu_events.match_found_no_sk_in_ct),
+			   (unsigned long long)
+			   	atomic64_read(&qtu_events.match_no_sk),
+			   (unsigned long long)
+			   	atomic64_read(&qtu_events.match_no_sk_file));
 
 		/* Count the following as part of the last item_index */
 		prdebug_full_state(0, "proc ctrl");
